Move zip from Zippable to BoxedZippable
Apparently GHC HEAD can identify type class function that can
never be used in an instance of the type class because of type
constraints. E.g. `zip` can never be used with `String` as a
target collection. `zip` is no in another type class that
automatically has an instance for every `Sequential` collection
type that can store tuples.